There were some changes that have fallen out in the update. Those were in the fields and tabs classes.
I have modified the tax classes according to the payment method.
this because the *_vend.php extend the normal customer class.

in the journal i have added a check-box so you can switch the auto_complete function on and of. 

In the config.php i have switched two items of there rank this made more sense. ( first the normal page then the manager below it.)