Loads of updates,
1 i have written a error reporting function this will catch all triggered errors and place them in a error file located in my_files/my_company/error.log. If it is a non recoverable error it will show a appropriate message on screen preventing the "white pages". The error log should help us with development. 

2 modified some parts to create a ajax error page instead of the current solution. ( if there is a sql error you will get a normal pop_up and if the you are logged out you will get the message saying that.)

3 added the .htacces files for security.

4 removed a lot of duplicate defined keys from the language files.

5 Changed the splitPageResults function so that it doesn't Execute the same sql statement again.
now the splitPageResults must be Executed after the sql ( this is diffent.) and the sql containing the results must contain SQL_CALC_FOUND_ROWS. Also the input for the functions have been changed makeing it simpler.

6 the phreepos module now has tax per till. 

7 in the inventory module when a new image is added it will check if the image name is already in the database. preventing accedently overwriding the previous image.

8 in all modules i have changed the prerequisites and version to doubles without quotes.
I have removed the quotes because this will force us to make it into doubles or integers.
There were a few that had two point making it into a string and then eg( '3.7.1' > '3.8' would return true) the upgrading would go wrong on next release.